Fix VirtualBox shared folder detection when using boot2docker
When initializing boot2docker, `check_for_shared_folders` is being
called before the `DOCKER_HOST_SSH_COMMAND` variable, which it uses, is
intialized. This has the effect of running commands locally instead of on
the boot2docker instance and, in this case, silently failing to detect
shared folders.

Moving `check_for_shared_folders` to the end of `init_boot2docker` fixes
this.

This seems to have been caused by the shift from using `VBoxManage` for
shared folder detection to `boot2docker ssh`, and then to the
`DOCKER_HOST_SSH_COMMAND` variable, without changing when
`check_for_shared_folders` is run.